chore: move JacksonTypeScanner to new package
