+++ sda-commons-server-dropwizard/src/main/java/org/sdase/commons/server/dropwizard/bundles/scanner/JacksonTypeScanner.java
@@ -1,4 +1,4 @@
-package org.sdase.commons.server.dropwizard.bundles;
+package org.sdase.commons.server.dropwizard.bundles.scanner;
 
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.databind.BeanDescription;
@@ -14,8 +14,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -26,7 +24,7 @@ public class JacksonTypeScanner {
       Set.of(String.class, File.class, Duration.class, DataSize.class);
 
   private static final Logger LOG = LoggerFactory.getLogger(JacksonTypeScanner.class);
-  public static final String MAP_STRING_KEY_NAME = "<key>";
+  static final String MAP_STRING_KEY_NAME = "<key>";
 
   private final ObjectMapper mapper;
   private final Set<Class<?>> plainTypes;
@@ -125,112 +123,5 @@ public class JacksonTypeScanner {
       }
     }
     return primaryType;
-  }
-
-  public static class MappableField {
-    private final List<String> jsonPathToProperty;
-    private final Type propertyType;
-    private final String environmentVariableName;
-    private final String propertyTypeSimpleName;
-
-    public MappableField(List<String> jsonPathToProperty, Type propertyType) {
-      this.jsonPathToProperty = jsonPathToProperty;
-      this.propertyType = propertyType;
-      this.environmentVariableName =
-          jsonPathToProperty.stream().map(String::toUpperCase).collect(Collectors.joining("_"));
-      this.propertyTypeSimpleName =
-          (propertyType instanceof Class)
-              ? ((Class<?>) propertyType).getSimpleName()
-              : propertyType.getTypeName();
-    }
-
-    private MappableField(
-        List<String> jsonPathToProperty, String environmentVariableName, Type propertyType) {
-      this.jsonPathToProperty = jsonPathToProperty;
-      this.propertyType = propertyType;
-      this.environmentVariableName = environmentVariableName;
-      this.propertyTypeSimpleName =
-          (propertyType instanceof Class)
-              ? ((Class<?>) propertyType).getSimpleName()
-              : propertyType.getTypeName();
-    }
-
-    public List<String> getJsonPathToProperty() {
-      return jsonPathToProperty;
-    }
-
-    public Type getPropertyType() {
-      return propertyType;
-    }
-
-    public String getPropertyTypeSimpleName() {
-      return propertyTypeSimpleName;
-    }
-
-    public String createTypeDescription() {
-      if (propertyType instanceof Class) {
-        var clazz = (Class<?>) propertyType;
-        if (clazz.isEnum()) {
-          return "enum " + getPropertyTypeSimpleName();
-        }
-        if (Map.class.isAssignableFrom(clazz)) {
-          return "Map [not supported]";
-        }
-      }
-      return getPropertyTypeSimpleName();
-    }
-
-    public String getEnvironmentVariableName() {
-      return environmentVariableName;
-    }
-
-    /**
-     * Resolves {@code MappableFields} from {@link #getEnvironmentVariableName()} that are available
-     * in the current context, considering variable parts like {@value
-     * JacksonTypeScanner#MAP_STRING_KEY_NAME}.
-     *
-     * @param availableKeysInContext the keys that are defined in the current context (e.g.
-     *     environment variable names)
-     * @return the {@code MappableField}s that are available in the current context, may be empty.
-     *     <strong>Note that the type is just copied and might be wrong for the actual
-     *     property.</strong>
-     */
-    public List<MappableField> expand(Set<String> availableKeysInContext) {
-      if (availableKeysInContext.contains(getEnvironmentVariableName())) {
-        return List.of(new MappableField(this.jsonPathToProperty, this.propertyType));
-      }
-      if (getEnvironmentVariableName()
-          .contains(JacksonTypeScanner.MAP_STRING_KEY_NAME.toUpperCase())) {
-        Pattern expandPattern = createExpandPattern(getEnvironmentVariableName());
-        return availableKeysInContext.stream()
-            .map(expandPattern::matcher)
-            .filter(Matcher::matches)
-            .map(
-                matcher -> {
-                  var newEnvironmentVariableName = matcher.group(0);
-                  var newPath = new ArrayList<>(getJsonPathToProperty());
-                  for (int i = 0; i < matcher.groupCount(); i++) {
-                    var realKeyPart = matcher.group(i + 1);
-                    for (int j = 0; j < newPath.size(); j++) {
-                      if (JacksonTypeScanner.MAP_STRING_KEY_NAME.equals(newPath.get(j))) {
-                        newPath.set(j, realKeyPart);
-                        break;
-                      }
-                    }
-                  }
-                  return new MappableField(newPath, newEnvironmentVariableName, getPropertyType());
-                })
-            .collect(Collectors.toList());
-      }
-      return List.of();
-    }
-
-    private Pattern createExpandPattern(String environmentVariableName) {
-      return Pattern.compile(
-          "^"
-              + environmentVariableName.replace(
-                  JacksonTypeScanner.MAP_STRING_KEY_NAME.toUpperCase(), "(.*)")
-              + "$");
-    }
   }
 }
