+++ sda-commons-server-dropwizard/src/test/java/org/sdase/commons/server/dropwizard/bundles/scanner/MappableFieldTest.java
@@ -1,4 +1,4 @@
-package org.sdase.commons.server.dropwizard.bundles;
+package org.sdase.commons.server.dropwizard.bundles.scanner;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.tuple;
@@ -7,7 +7,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import org.junit.jupiter.api.Test;
-import org.sdase.commons.server.dropwizard.bundles.JacksonTypeScanner.MappableField;
 
 class MappableFieldTest {
